Implement Fullscreen mode
A new menuitem toggles Fullscreen mode, which means that the
menubar is hidden, the window is maximized and window manager
decoration is removed. A new button appears in the navibar
which reverts this process.

Fullscreen relies entirely on the window manager, which has the
advantage that window manager keybindings have the same effect
as activating the Fullscreen menuitem. At the same time this
means that Fullscreen won't work at all without it.