Update interface and web featureswith preference changes.
The toolbar style, small toolbar, New Tab, Web Search and
Trash buttons are shown or hidden respectively as
the preference values change.
This is also an attempt at updating the web view's settings
immediately. Although it seems to work, an ugly warning is
emitted to the console on every change.