Implement togglable ticks in the extensions panel
Whether an extension is active or not is now reflected by a tick
mark that can be toggled with a mouse. Double clicking and
pressing space do the same. The buttons are thusly removed.
While we are at it, we also listen for status changes that might be
coming from a different window.
Sensitivity does no longer relate to activation status.