Make sure speed dial is shown in blank tabs
The speed dial is now mandatory, but it's still possible to get
a blank tab. With an empty home page and a displayed speed dial
on a new tab, going to any page and coming back leads to a blank page.

Work around that by checking whether the view is empty and setting the
uri to an empty string in this case, to hit the code path which makes
the speed dial kick in.