Initialize session in app/ private before adding URIs
If the session is not prepared properly URIs loading of
the first page(s) breaks. Also it is a good idea to sort
out all settings before loading anything to begin with.

You will now see "midori_load_soup_session was not called!" if
any pages try to load before the session is prepared.

Fixes: https://bugs.launchpad.net/midori/+bug/814268