Don't silently discard extensions which don't load
Broken extensions result in an according description and a
warning which aborts the test run.
Counting the number of expected extensions needs a test case;
a failed attempt is included together with a FIXME.

Also force katze_net_load_uri to be compiled in even if unused in
the core so that extensions (Feed Panel) can use it.

Fixes: https://bugs.launchpad.net/midori/+bug/1095194