Improving error message when gnome-keyring is not running.
Currently, when Midori cannot find a proper instance of gnome-keyring
running, it won't accept a self-signed certificate and will issue a
cryptic error message:

  "Error granting trust: Couldn't find a place to store the pinned certificate
   pkcs11:library-manufacturer=GNOME%20Keyring ,
   pkcs11:library-description=PKCS%2311%20Kit%20Trust%20Module"

This error is already explained in the FAQ, but we could do better: we
can provide a direct link to the FAQ section that teaches the user how
to fix this.  That's what we do in this commit.

As a reference, Debian bug #799963 started this discussion.