Implement panel in the browser window (#13)
![screenshot from 2018-08-21 23-51-20](https://user-images.githubusercontent.com/1204189/44431228-7f5c3d00-a59d-11e8-8872-9bddc1e7adc4.png)
A Gtk.Paned splits the browser window into what's considered the panel and the web view. Compared to the old code, the navigationbar is inside the paned and a Gtk.StackSwitcher (in a second headerbar)/ Gtk.Stack takes the role of the toolbar/ notebook.

Nothing implements a panel right now, although it's exposed via peas.