Update/ markdownify README(.md) (#29)
This change consolidates README and HACKING into an updated README.md with GitHub-compatible Markdown. Tweaks to the description which is re-used in snap and appstream descriptions are also included as well as a link to the new Telegram group.

Note: The instructions for Windows with the latest sources are currently untested.