Show tab popover relative to urlbar with no opacity changes (#164)
Popovers really don't work well unless they're given an explicit
position. So for a tab that may be the tab itself or better yet
the urlbar of the containing browser.

With some themes the entire tabs goes blank due to opacity
changes, and even if it doesn't the translucency may leave
the user unclear as to what's happening. So let's not do this.

Fixes: #163

Also use the popover to re-implement permission requests.

Fixes: #158