Add Trust button to the security details popover
If a certificate is available but considered invalid, the Trust button
shows up in the details popover accessible via the icon in the urlbar.
A temporary exception which is valid for the remainder of the session
is added for the specific certificate and host combination.

Note: No attempt is made to preserve the exception.

Fixes: #182